requirements.txt
httpx>=0.27
aiosqlite>=0.20
python-dotenv>=1.0
tenacity>=8.2
